feat(website): add `TextEffect` component
