+++ app/package.json
@@ -17,8 +17,8 @@
     "@mdx-js/loader": "^3.1.0",
     "@mdx-js/react": "^3.1.0",
     "@next/mdx": "^15.3.1",
-    "framer-motion": "12.0.0-alpha.2",
     "highlight.js": "^11.11.1",
+    "motion": "^12.18.1",
     "next": "15.2.4",
     "react": "19.1.0",
     "react-dom": "19.1.0",
