+++ app/src/lib/@types/types.ts
@@ -1,1 +1,5 @@
 export type ArgsFunction = (...args: any) => void
+
+export type PresetType = "blur" | "fade-in-blur" | "scale" | "fade" | "slide"
+
+export type PerType = "word" | "char" | "line"
