+++ app/src/ui/header/header.tsx
@@ -3,7 +3,7 @@ import Link from "next/link"
 import Image from "next/image"
 import { useEffect, useState } from "react"
 import { HeaderMenu } from "./header-menu"
-import { AnimatePresence } from "framer-motion"
+import { AnimatePresence } from "motion/react"
 import { merge } from "@halvaradop/ui-core"
 import menu from "@/assets/menu.svg"
 import arrow from "@/assets/arrow.svg"
