+++ app/src/ui/text-effect.tsx
@@ -1,145 +1,15 @@
 "use client"
-import { merge } from "@halvaradop/ui-core"
 import { AnimatePresence, motion } from "motion/react"
 import type { TargetAndTransition, Transition, Variant, Variants } from "motion/react"
-import React from "react"
-
-export type PresetType = "blur" | "fade-in-blur" | "scale" | "fade" | "slide"
-
-export type PerType = "word" | "char" | "line"
-
-export type TextEffectProps = {
-    children: string
-    per?: PerType
-    as?: keyof React.JSX.IntrinsicElements
-    variants?: {
-        container?: Variants
-        item?: Variants
-    }
-    className?: string
-    preset?: PresetType
-    delay?: number
-    speedReveal?: number
-    speedSegment?: number
-    trigger?: boolean
-    onAnimationComplete?: () => void
-    onAnimationStart?: () => void
-    segmentWrapperClassName?: string
-    containerTransition?: Transition
-    segmentTransition?: Transition
-    style?: React.CSSProperties
-}
-
-const defaultStaggerTimes: Record<PerType, number> = {
-    char: 0.03,
-    word: 0.05,
-    line: 0.1,
-}
-
-const defaultContainerVariants: Variants = {
-    hidden: { opacity: 0 },
-    visible: {
-        opacity: 1,
-        transition: {
-            staggerChildren: 0.05,
-        },
-    },
-    exit: {
-        transition: { staggerChildren: 0.05, staggerDirection: -1 },
-    },
-}
-
-const defaultItemVariants: Variants = {
-    hidden: { opacity: 0 },
-    visible: {
-        opacity: 1,
-    },
-    exit: { opacity: 0 },
-}
-
-const presetVariants: Record<PresetType, { container: Variants; item: Variants }> = {
-    blur: {
-        container: defaultContainerVariants,
-        item: {
-            hidden: { opacity: 0, filter: "blur(12px)" },
-            visible: { opacity: 1, filter: "blur(0px)" },
-            exit: { opacity: 0, filter: "blur(12px)" },
-        },
-    },
-    "fade-in-blur": {
-        container: defaultContainerVariants,
-        item: {
-            hidden: { opacity: 0, y: 20, filter: "blur(12px)" },
-            visible: { opacity: 1, y: 0, filter: "blur(0px)" },
-            exit: { opacity: 0, y: 20, filter: "blur(12px)" },
-        },
-    },
-    scale: {
-        container: defaultContainerVariants,
-        item: {
-            hidden: { opacity: 0, scale: 0 },
-            visible: { opacity: 1, scale: 1 },
-            exit: { opacity: 0, scale: 0 },
-        },
-    },
-    fade: {
-        container: defaultContainerVariants,
-        item: {
-            hidden: { opacity: 0 },
-            visible: { opacity: 1 },
-            exit: { opacity: 0 },
-        },
-    },
-    slide: {
-        container: defaultContainerVariants,
-        item: {
-            hidden: { opacity: 0, y: 20 },
-            visible: { opacity: 1, y: 0 },
-            exit: { opacity: 0, y: 20 },
-        },
-    },
-}
-
-const AnimationComponent: React.FC<{
-    segment: string
-    variants: Variants
-    per: "line" | "word" | "char"
-    segmentWrapperClassName?: string
-}> = React.memo(({ segment, variants, per, segmentWrapperClassName }) => {
-    const content =
-        per === "line" ? (
-            <motion.span variants={variants} className="block">
-                {segment}
-            </motion.span>
-        ) : per === "word" ? (
-            <motion.span aria-hidden="true" variants={variants} className="inline-block whitespace-pre">
-                {segment}
-            </motion.span>
-        ) : (
-            <motion.span className="inline-block whitespace-pre">
-                {segment.split("").map((char, charIndex) => (
-                    <motion.span
-                        key={`char-${charIndex}`}
-                        aria-hidden="true"
-                        variants={variants}
-                        className="inline-block whitespace-pre"
-                    >
-                        {char}
-                    </motion.span>
-                ))}
-            </motion.span>
-        )
-
-    if (!segmentWrapperClassName) {
-        return content
-    }
-
-    const defaultWrapperClassName = per === "line" ? "block" : "inline-block"
-
-    return <span className={cn(defaultWrapperClassName, segmentWrapperClassName)}>{content}</span>
-})
-
-AnimationComponent.displayName = "AnimationComponent"
+import { TextEffectProps } from "@/lib/@types/props"
+import type { PerType } from "@/lib/@types/types"
+import {
+    defaultContainerVariants,
+    defaultItemVariants,
+    defaultStaggerTimes,
+    presetVariants,
+} from "@/ui/motion/text-effect.motion"
+import { AnimationComponent } from "@/ui/animation-component"
 
 const splitText = (text: string, per: PerType) => {
     if (per === "line") return text.split("\n")
